datapath: move Orchestrator to pkg/endpoint/types.Orchestrator
